refactor: change structure of page Register
